Update lwotai.py
recruitChoice was incorrectly processing the 2nd bubble of the flowchart.  The logic
was checking for cells < 2*country resources instead of cells < 2*ops.  There was also
a similar bug here that was assigning the first country AFG a zero score value which 
was causing it to get incorrectly chosen as a recruit choice.

Cards 87,88,89 had a simple typo/bug that was causing the script to crash when 
those card events were played as the US player.